requirements.txt
fastapi==0.75.2
pydantic==1.7.3
PyJWT==2.3.0
uvicorn==0.14.0
